Basic setup of JFrame
Put together the basic format of the JFrame such as size, layout, etc. Decided to switch between panels when a button is pressed in order to perform the appropriate functions.  Also, Added a skeleton for the Action Listener that will be used for the main page/panel and its 6 buttons. Added the first button to the main panel which allows user to view panel. It is in the beginning stages so it doen't do anything so far.